Don’t try to parse the response, just pass it on.
We don’t need to actually do anything with the data so it’s pointless having something like Jackson attempt to transform the data. Instead we just pass it on. We shouldn’t be reading this into memory and should really be streaming it between request and response, but it will do for now.