CM-50 Add manual refresh of access tokens
This isn’t ideal and in the future we should have the RestTemplate/WebClient handle this transparently, but for now it’s good enough.

We should also handle HTTP status codes as well so that when Canvas says it’s an invalid token we try once to refresh the access token. But again, this can wait.