CM-112 Use the same codepath when token fails
Rather than handling the failure to refresh the access token differently, we just let it fall through and attempt to use the access token which we expect will fail. This way the error from renewing a access token failing and using a invalid access token will be the same and the client side code has one less case to handle and test for.