CM-179 Correctly handle concurrent token renewals
Now that we have multiple concurrent request being made when the Course Management tool is launched it highlights a problem that we may start two access token renewals concurrently. There is no grace in Canvas for having old tokens continue to work, so after a token is renewed the old token stops working immediately.

So we lock a user’s token in the DB before we start a token refresh, and then do the token refresh. This will block other threads, and they should end up just getting the newly refreshed token from the DB.

Doing the locking in the DB should allow for the application to scale across multiple nodes if we need to in the future.