AB#57430 Allow for access token to be refreshed
The best way to check to see if we still have a valid token is to attempt to use the refresh token to get a new access token. This allows the client to make an authenticated request and the proxy will attempt to renew the access token if it is out of date. The URL to use is /token/refresh and if there is a non-expired access token it will return 200, otherwise it will either redirect (3xx) to have a token granted, or return unauthorized (403).

It accepts the parameter `force=true` which when set ignores the expiry time of the access token and renews it anyway.