AB#57430 Don't use session on token check
We set the cookies used for the grant process to be just for that path so that it's not used when checking to see if the current user has a token. Otherwise we don't get an error when an expired JWT is used to check as if the user has a cookie that is used to establish the session instead.