AB#69768 Allow JWT authentication to persist
When the OAuth flow to grant permission happens we need a cookie based session so that we can associate the initial request with the completed grant.

In the longer run we could probably see if we can actually remove cookies and encode the information needed for the return in the URL we pass across.