AB#65252 Merge LTI Service into Proxy
This has both the LTI Service and the Proxy in the same webapp.

This uses `@Qualifiers` to keep the LTI and the Proxy components apart
when both services want a bean of a particular type. The packages of the
original services have been kept the same and anything that is common to
both services is in the `.ctl` package.