AB#65252 Support multiple config prefixes
The problem is that in the proxy config we would have a config value of:

    spring.security.oauth2.client.registration.universityofoxford-sa.client-name=...

and in the lti config we would have the same:

    spring.security.oauth2.client.registration.universityofoxford-sa.client-name=...

So in merging the 2 config we would have to change `universityofoxford-sa` key and this is in the Canvas configuration which means updating every tool.

Rather than having to edit lots of config and make sure everything still matches up, this allows the LTI and Proxy configuration to be imported at the same time under different prefixes. This way we just have to edit the prefix when we merge the configuration and don't have to change lots of registration names.