AB#65252 We weren't getting scopes on access tokens
When we were persisting the access tokens we weren't saving the scopes as we weren't using our custom OAuth client setup on the refresh endpoint.

This meant that when we checked that the scopes on the token were the same as those in the registration it would always fail and get the user to re-authenticate.

This is needed because of a fix for https://spring.io/security/cve-2022-31690 which doesn't matter for us (but we need scopes to detect when client needs to be re-authorized).