AB#65264 Add example sections that are useful.
These examples are the minimum values needed when configuring a tool. Although the API support is much more flexible we

This is needed because the JSON de-serializer always uses the newer CLIENT_SECRET_POST instead of POST when de-serializing JSON, this means that when I updated a client registration it stopped working.

We could have updated the de-serialization code so that it used POST instead, but we are going to have to move toe CLIENT_SECRET_POST at some point so we may have well added support for it now.