AB#65264 Handle HTTP 302 errors
When getting an error back from the token endpoint this code handles the HTTP 302 error correctly and extracts out the error and error description. This will no longer be supported in a short time, but until then it makes debugging configuration errors much easier as we propogate the error back to the frontend, instead of having to use a debugger.