AB#65422 Handle tokens responses of HTTP 401
When we have the wrong secret set then Canvas returns a HTTP 401 with a useful message in the body. We should throw this error to make debugging easier.

We also attempt to parse the body before looking at the header as there's more information in the body in the case of Canvas.