Unstable local test fix
When attempting to do a release locally I was getting test failures, this was because it was running the code too fast and generating the same timestamp for issued at and expiry.

Adding an hour to the expiry time (which is closer to how it would work in practice) fixes this.