AB#70623 Allow a standard login initiation URL.
At the start of the LTI launch Canvas sends a request to the login initiation URL, we used to require that it include the registration ID in the URL so that we could lookup which tool was being launched. This meant that you had to manage a unique namespace of registration IDs and these had to be known before you could fully configure the tool in Canvas.

This change means we can work out which tool is being launched from the requests parameters (specifically the client ID) and so we don't need to include registration ID in the initiation URL (although that method still works).

While this does make the service slightly more complicated it means configuration is much simpler and that you can configure Canvas completely and then configure tool-support afterwards.