AB#75560 Ensure that we correctly handle sessions.
We weren't correctly setting up our session which resulted in the session from one user being used to store tokens from a second user.

To reproduce this if you use an application as one user and grant it access to your Canvas account and then use the same application as a different Canvas user it should correctly save the tokens for the different user (and not update the first user).

Also bumped up the Spring Boot version to pull in newer Spring Security versions.