AB#78964 Add support for anonymous LTI launches.
This also adds support to the API proxy side to handle tokens from an anonymous LTI launch and just pass the request through to Canvas without any Bearer token on it.

This is useful because we set CORS headers on our responses which allows these API requests to be made from domains other than the Canvas domain.

Most of the change for this work is in the Spring Security LTI13 library.