AB#79264 Handle the case where the filter doesn't run.
If the filter doesn't run (normally because the original request can't be found) then the OAuthErrorController gets called but there's no error.

I don't think this is anything that we are doing wrong, but is more likely to be browser retries due to a poor network connection. I can't see a simple way we could recover from this so the best solution looks to be to display a helpful message to the end user.