AB#81478 Issue JWTs under the correct name.
All LTI JWTs that we issue should be for https://lti(-dev).canvas.ox.ac.uk as this is what all the consumers are expecting (ie the tools are configured to only accept that issuer). The problem was that it was signing tokens under the new domain name.

While I could have duplicated the config in the EB environment I wanted to make it clear what the values are set to and it doesn't involve that much duplication.