AB#86377 Use template to reference solution stack
We have managed updates enabled for the EB environment which results in the running environment automatically being updated to the latest version shortly after it's released. We have the solution stack parameter in the templates so that when creating a new environment it selects the newest one, but we don't want changes in the parameter to cause the stack to re-deploy (as it will fail). So we use a template which means the solution stack parameter is only referenced at environment creation time.